+++ tests/unit/test_variants.c
@@ -947,27 +947,37 @@ static void test_xxh64_and_xxh3_64_differ_for_same_input(void)
 
 static void test_xxh3_64_update_null_state_returns_error(void)
 {
+#ifndef NDEBUG
     TEST_ASSERT_EQUAL_INT(XXH3_ERROR, xxh3_64_update(NULL, LOREM, strlen(LOREM)));
+#endif
 }
 
 static void test_xxh3_64_digest_null_state_returns_zero(void)
 {
+#ifndef NDEBUG
     TEST_ASSERT_EQUAL_UINT64(0, xxh3_64_digest(NULL));
+#endif
 }
 
 static void test_xxh3_128_update_null_state_returns_error(void)
 {
+#ifndef NDEBUG
     TEST_ASSERT_EQUAL_INT(XXH3_ERROR, xxh3_128_update(NULL, LOREM, strlen(LOREM)));
+#endif
 }
 
 static void test_xxh32_update_null_state_returns_error(void)
 {
+#ifndef NDEBUG
     TEST_ASSERT_EQUAL_INT(XXH3_ERROR, xxh32_update(NULL, LOREM, strlen(LOREM)));
+#endif
 }
 
 static void test_xxh64_update_null_state_returns_error(void)
 {
+#ifndef NDEBUG
     TEST_ASSERT_EQUAL_INT(XXH3_ERROR, xxh64_update(NULL, LOREM, strlen(LOREM)));
+#endif
 }
 
 /* ----------------------------------------------------------- main test runner */